+++ game-frontend/projects/game-frontend/src/app/modules/ranking/types/ranking-entry.ts
@@ -6,6 +6,8 @@
  * @export
  * @interface RankingEntry
  */
+import {CommonEntity} from '@owge/core';
+
 export interface RankingEntry {
     position: number;
     points: number;
@@ -13,4 +15,7 @@ export interface RankingEntry {
     username: string;
     allianceId: number;
     allianceName: string;
+    faction: {
+        name: string; description: string; imageUrl: string;
+    };
 }